requirements.txt
aiohttp
python_socketio
picamera
